Make `rosetta-rpc` depend on `near-network-primitives` instead of `near-network`
