+++ chain/client/src/tests/query_client.rs
@@ -9,9 +9,9 @@ use near_actix_test_utils::run_actix;
 use near_crypto::{InMemorySigner, KeyType};
 use near_logger_utils::init_test_logger;
 use near_network::test_utils::MockPeerManagerAdapter;
-use near_network::types::{NetworkClientMessages, NetworkClientResponses};
 use near_network_primitives::types::{
-    NetworkViewClientMessages, NetworkViewClientResponses, PeerInfo,
+    NetworkClientMessages, NetworkClientResponses, NetworkViewClientMessages,
+    NetworkViewClientResponses, PeerInfo,
 };
 use near_primitives::block::{Block, BlockHeader};
 use near_primitives::time::Utc;
