+++ integration-tests/src/tests/client/chunks_management.rs
@@ -17,8 +17,10 @@ use near_client::test_utils::setup_mock_all_validators;
 use near_client::{ClientActor, GetBlock, ViewClientActor};
 use near_logger_utils::init_test_logger;
 use near_network::types::PeerManagerMessageRequest;
-use near_network::types::{NetworkClientMessages, NetworkRequests, NetworkResponses};
-use near_network_primitives::types::{AccountIdOrPeerTrackingShard, PeerInfo};
+use near_network::types::{NetworkRequests, NetworkResponses};
+use near_network_primitives::types::{
+    AccountIdOrPeerTrackingShard, NetworkClientMessages, PeerInfo,
+};
 use near_primitives::hash::CryptoHash;
 use near_primitives::transaction::SignedTransaction;
 use near_primitives::types::AccountId;
