+++ integration-tests/src/tests/client/process_blocks.rs
@@ -26,11 +26,12 @@ use near_client::{Client, GetBlock, GetBlockWithMerkleTree};
 use near_crypto::{InMemorySigner, KeyType, PublicKey, Signature, Signer};
 use near_logger_utils::init_test_logger;
 use near_network::test_utils::{wait_or_panic, MockPeerManagerAdapter};
-use near_network::types::{
-    FullPeerInfo, NetworkClientMessages, NetworkClientResponses, NetworkRequests, NetworkResponses,
+use near_network::types::{NetworkRequests, NetworkResponses};
+use near_network::types::{PeerManagerMessageRequest, PeerManagerMessageResponse};
+use near_network_primitives::types::{
+    FullPeerInfo, NetworkClientMessages, NetworkClientResponses, NetworkInfo, PeerChainInfoV2,
+    PeerInfo, ReasonForBan,
 };
-use near_network::types::{NetworkInfo, PeerManagerMessageRequest, PeerManagerMessageResponse};
-use near_network_primitives::types::{PeerChainInfoV2, PeerInfo, ReasonForBan};
 use near_primitives::block::{Approval, ApprovalInner};
 use near_primitives::block_header::BlockHeader;
 use near_primitives::epoch_manager::RngSeed;
