+++ integration-tests/src/tests/network/stress_network.rs
@@ -15,8 +15,8 @@ use near_network::routing::start_routing_table_actor;
 use near_network::test_utils::{
     convert_boot_nodes, open_port, GetInfo, StopSignal, WaitOrTimeoutActor,
 };
-use near_network::types::NetworkClientResponses;
 use near_network::PeerManagerActor;
+use near_network_primitives::types::NetworkClientResponses;
 use near_network_primitives::types::{
     NetworkConfig, NetworkViewClientMessages, NetworkViewClientResponses,
 };
